[IMP] Allow to correctly set default values for selectable fields
Most of the time it works nicely, but for the case when the formio
selectable field is fetching the options from Odoo, it's not possible
to correctly set the value as expected by formio (in general an object
with the shame {id: int: label: string}).

This PR allows to do that, and at the same time, makes possible to
include more data in the options values that can be used in the option
template or used by other field's logic.

So correct configuration for /data url selectable fields could be:
user_field; {id: partner_id.id, label: partner_id.name}

but even could be:
{id: partner_id.id, label: partner_id.name, country: partner_id.country_id.name}