[IMP] Allow to use complex filtering on formio selectable fetched options
This PR include a whole now set of operators and query syntax for
formio selectable fields that fetched their values form Odoo by using
the /data endpoint as explained in documentation.

It also applies values conversion. That means a value of 'True' or 'False'
will be converted to True of False and correctly applied to the search
domain.
The same for integers and float (included external ids of related models)
that will be converted and used correctly in the odoo domain.

The new syntax is inspired in FIQL, and the mapping to Odoo domain
operators follow this mapping:

OPERATORS_MAPPING = {
    '==': '=',
    '=ne=': '!=',
    '=gt=': '>',
    '=ge=': '>=',
    '=lt=': '<',
    '=le=': '<=',
    '=like=': 'like',
    '=nlike=': 'not like',
    '=ilike=': 'ilike',
    '=nilike=': 'not ilike',
    '=in=': 'in',
    '=nin=': 'not in',
    '=child=': 'child_of',
    '=parent=': 'parent_of',
}

So using the operators in the left, we could do now in formio:

model=res.partner&label=name&domain=id=lt=100;name=like="arl";parent_id==False;customer_rank=ge=50