[IMP] formio: Workaround Safari with unusable Columns in a Data Grid.
Workaround for Safari 16.6 (AFAIK), with Columns in a Data Grid.
The Columns become too narrow and unusable in a Data Grid.
GitHub issue: https://github.com/formio/formio.js/issues/5426

This workaround adds 1 CSS rule, which adds a `width: 100%` to the
first `<td/>` (column) targeted in a Data Grid table.

    width: 100%;
}