[IMP] formio: Improve the formio.js library registration to allow only registered versions.
Improve the formio.js library registration (downloader, importer) with
a new setting to allow only registered versions.

This adds a new system parameter `ir.config_parameter` which currently
defaults to 'v4' and can be modified in the configuration window.

The allowed setting is a comma separated string (list) of formio.js
versions to register. Examples:
- v4,v5
- v4.17,v4.18